backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
rembg==2.0.56
pillow==10.3.0
python-multipart==0.0.9
